requirements.txt
pybithumb>=0.2.8
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
